requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.3.0
asgiref==3.4.1
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.6.0
fastapi==0.68.0
fastapi-mail==0.4.1
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
redis==3.5.3
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
